WIP: try to split in 2 subpackages
