+++ projects/akita-filters-plugin/src/lib/datasource/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of akita-mat-data-source
+ */
+
+
+
+export * from './akita-mat-data-source';
