+++ projects/akita-filters-plugin/src/lib/plugin/akita-filters-plugin.ts
@@ -1,5 +1,5 @@
-import {AkitaFilter, AkitaFiltersStore, createFilter, FiltersState} from './akita-filters-store';
-import {AkitaFiltersQuery} from './akita-filters-query';
+import {AkitaFilter, AkitaFiltersStore, createFilter, FiltersState} from 'akita-filters/plugin';
+import {AkitaFiltersQuery} from 'akita-filters/plugin';
 import {combineLatest, isObservable, merge, Observable, ObservedValueOf, of} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {
