+++ projects/akita-filters-plugin/src/lib/plugin/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "akita-filters/plugin",
+  "ngPackage": {
+    "lib": {
+      "entryFile": "./public_api.ts"
+    }
+  }
+}
