+++ projects/akita-filters-plugin/src/lib/plugin/public_api.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of akita-filters-plugin
+ */
+
+export * from './akita-filters-plugin';
+export * from './akita-filters-query';
+export * from './akita-filters-store';
+export * from './filters-utils';
