+++ projects/akita-filters-plugin/src/public_api.ts
@@ -2,9 +2,7 @@
  * Public API Surface of akita-filters-plugin
  */
 
-export * from './lib/akita-filters-plugin';
-export * from './lib/akita-filters-query';
-export * from './lib/akita-filters-store';
-export * from './lib/filters-utils';
-
-export * from './lib/bonus/akita-mat-data-source';
+export * from './lib/plugin/akita-filters-plugin';
+export * from './lib/plugin/akita-filters-query';
+export * from './lib/plugin/akita-filters-store';
+export * from './lib/plugin/filters-utils';
